+++ dbcrossbar/tests/cli/cp/combined.rs
@@ -2,11 +2,7 @@
 
 use cli_test_dir::*;
 use difference::assert_diff;
-use opinionated_telemetry::{
-    current_span_as_env, set_parent_span_from_env, AppType, TelemetryConfig,
-};
 use std::{fs, process::Stdio};
-use tracing::info_span;
 
 use super::*;
 
@@ -23,21 +19,9 @@ fn cp_csv_to_postgres_to_gs_to_csv() {
     let gs_dir_2 = gs_test_dir_url("cp_csv_to_postgres_to_gs_to_csv_2");
     let pg_table_2 = post_test_table_url("cp_csv_to_postgres_to_gs_to_csv_2");
 
-    // Just for fun, set up a trace across multiple calls to `dbcrossbar`.
-    let telemetry_handle = TelemetryConfig::new(
-        AppType::Cli,
-        env!("CARGO_PKG_NAME"),
-        env!("CARGO_PKG_VERSION"),
-    )
-    .install_sync()
-    .expect("could not install telemetry");
-    let span = info_span!("cp_csv_to_postgres_to_gs_to_csv").entered();
-    set_parent_span_from_env();
-
     // CSV to Postgres.
     testdir
         .cmd()
-        .envs(current_span_as_env())
         .args([
             "cp",
             "--if-exists=overwrite",
@@ -53,7 +37,6 @@ fn cp_csv_to_postgres_to_gs_to_csv() {
     // messing up later tests.)
     testdir
         .cmd()
-        .envs(current_span_as_env())
         .args(["schema", "conv", &pg_table, "postgres-sql:pg.sql"])
         .stdout(Stdio::piped())
         .tee_output()
@@ -67,7 +50,6 @@ fn cp_csv_to_postgres_to_gs_to_csv() {
     // Postgres to gs://.
     testdir
         .cmd()
-        .envs(current_span_as_env())
         .args(["cp", "--if-exists=overwrite", &pg_table, &gs_dir])
         .tee_output()
         .expect_success();
@@ -75,7 +57,6 @@ fn cp_csv_to_postgres_to_gs_to_csv() {
     // gs:// to BigQuery.
     testdir
         .cmd()
-        .envs(current_span_as_env())
         .args([
             "cp",
             "--if-exists=overwrite",
@@ -89,7 +70,6 @@ fn cp_csv_to_postgres_to_gs_to_csv() {
     // BigQuery to gs://.
     testdir
         .cmd()
-        .envs(current_span_as_env())
         .args([
             "cp",
             "--if-exists=overwrite",
@@ -104,7 +84,6 @@ fn cp_csv_to_postgres_to_gs_to_csv() {
     // CSV file for the final comparison below.)
     testdir
         .cmd()
-        .envs(current_span_as_env())
         .args([
             "cp",
             "--if-exists=overwrite",
@@ -118,7 +97,6 @@ fn cp_csv_to_postgres_to_gs_to_csv() {
     // PostgreSQL back to CSV for the final comparison below.
     testdir
         .cmd()
-        .envs(current_span_as_env())
         .args([
             "cp",
             &format!("--schema=postgres-sql:{}", schema.display()),
@@ -133,9 +111,6 @@ fn cp_csv_to_postgres_to_gs_to_csv() {
         fs::read_to_string(testdir.path("out/cp_csv_to_postgres_to_gs_to_csv_2.csv"))
             .unwrap();
     assert_diff!(&expected, &actual, ",", 0);
-
-    drop(span);
-    telemetry_handle.flush_and_shutdown();
 }
 
 #[test]